Ajout affichage chart

+++ src/Controler/PriceNetworkService.java
@@ -50,7 +50,7 @@ public class PriceNetworkService extends Thread{
 				if( b[0] == '|')
 				{
 					System.out.println("[INFO]"+sValue);
-					netService.refreshValue(tick,sValue);
+					netService.refreshValue(tick,Float.valueOf(sValue));
 					tick++;
 					sValue = "";
 				}
